Make stuff better
```
name                                    old time/op  new time/op  delta
Contains1kX10kX5/contains-6              113ns ± 2%    90ns ± 1%  -20.16%  (p=0.000 n=9+8)
Contains1kX10kX5/containsHash-6         95.0ns ± 1%  75.8ns ± 2%  -20.26%  (p=0.000 n=9+9)
Contains100kX10BX20/contains-6          1.01µs ± 8%  0.17µs ± 1%  -82.75%  (p=0.000 n=9+9)
Contains100kX10BX20/containshash-6       969ns ± 1%   144ns ± 0%  -85.14%  (p=0.000 n=8+8)
Contains94percentMisses/contains-6       258ns ± 2%    91ns ± 1%  -64.79%  (p=0.000 n=9+8)
Contains94percentMisses/containsHash-6   238ns ± 2%    70ns ± 1%  -70.76%  (p=0.000 n=9+9)
```